requirements.txt
# Usage: pip install -r requirements.txt
python-docx>=1.1.0
pandas>=2.1.1